fix(vm): preserve int64/uint64 precision in arithmetic operations on 32-bit
On 32-bit architectures, arithmetic operations between int64 operands
silently truncated to int (32-bit), causing overflow. Fix the code
generator to promote result type to int64/uint64 when either operand
is int64/uint64.

Fixes #804

+++ expr_test.go
@@ -805,15 +805,15 @@ func TestExpr(t *testing.T) {
 		},
 		{
 			`Uint64 + 0`,
-			0,
+			uint64(0),
 		},
 		{
 			`Uint64 + Int64`,
-			0,
+			uint64(0),
 		},
 		{
 			`Int32 + Int64`,
-			0,
+			int64(0),
 		},
 		{
 			`Float64 + 0`,
